pyproject.toml
[project]
name = "ff4kb"
version = "1.0.0"
requires-python = ">=3.12"

[tool.poetry]
name = "ff4"
version = "1.0.0"
description = "ff4kb.aexoden.com"
authors = ["Jason Lynch <jason@aexoden.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
dj-database-url = "^2.2.0"
django-configurations = "^2.5"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
django4-bootstrap-breadcrumbs = "^0.10.0"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
preview = true
select = [
    "A",
    "ANN",
    "ARG",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "COM",
    "CPY",
    "DTZ",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FA",
    "FBT",
    "FIX",
    "FURB",
    "G",
    "I",
    "ICN",
    "INP",
    "ISC",
    "LOG",
    "N",
    "PERF",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLF",
    "SLOT",
    "T10",
    "TCH",
    "TD",
    "TID",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "COM812",
    "ISC001",
    "S404",
    "S603",
]


[tool.ruff.lint.flake8-copyright]
notice-rgx = "(?i)SPDX-FileCopyrightText: \\d{4}"

[tool.ruff.lint.isort]
lines-between-types = 1
